+++ src/pages/Authorization/SignIn/index.jsx
@@ -1,5 +1,5 @@
-import { SignInForm } from './components/SignInForm'
-import AuthPageTemplate from '../components/AuthPageTemplate';
+import SignInForm from './components/SignInForm'
+import AuthPageTemplate from '../components/AuthPageTemplate/AuthPageTemplate';
 
 const SignIn = () => {
   return (
