+++ src/pages/Authorization/SignUp/index.jsx
@@ -1,5 +1,5 @@
-import { SignUpForm } from './components/SignUpForm';
-import AuthPageTemplate from '../components/AuthPageTemplate';
+import SignUpForm from './components/SignUpForm';
+import AuthPageTemplate from '../components/AuthPageTemplate/AuthPageTemplate';
 
 const SignUp = props => {
   return (
